Refactor list orgs specs
