+++ src/components/ListOrgs.vue
@@ -11,7 +11,7 @@
         </div>
         <div class="text-md text-gray-500 ml-6">View organizations asssigned to your account.</div>
       </div>
-      <PvTabView v-if="claimsLoaded" v-model:active-index="activeIndex" lazy class="mb-7">
+      <PvTabView v-if="claimsLoaded" v-model:active-index="activeIndex" lazy class="mb-7" data-cy="orgs-list">
         <PvTabPanel v-for="orgType in orgHeaders" :key="orgType" :header="orgType.header">
           <div class="grid column-gap-3 mt-2">
             <div
